One-Two terms to finish
holder, held by : Context needed: does it refers to the owner of copyright (=titular de los derechos ) or the sponsor (=patrocinador, patrocinado por) / organizer (=organizador, organizado por) of an event or what else? 

For Spanish accented characters , " \' + vowel" is used for all vowels, including "í", i.e. as "\'i" instead of "\{i}" . This latter form appears in standard strings of file. I do not know if accented "í" must be so.